requirements.txt
firebase-admin==6.5.0
openai==1.37.0
openai-whisper==20231117
python-dotenv==1.0.1
yt-dlp==2024.7.25
